proje2/requirements.txt
flask
pyjwt
python-dotenv
flask-cors
sqlalchemy
werkzeug
dotenv
subprocess
